Added some comments and fixed a bug in mod_twitter

+++ modules/mod_twitter.py
@@ -19,35 +19,38 @@ from src.output import *
 from xml.dom import minidom
 import time, urllib
 
+#Read in code to dynamically get the module name.
 execfile('src/getname')
 
 #Get configuration options
 count = check_config("MOD_TWITTER_COUNT=")
-interval = float(check_config("MOD_TWITTER_INTERVAL="))
+interval = int(check_config("MOD_TWITTER_INTERVAL="))
 
 def main(query,*args):
 	id = 0
 	global count, interval
 	while True:
 
-	    url = "http://search.twitter.com/search.atom?rpp=" + count + "&q=%s&since_id=%s" % (query, id)
-	    xml = urllib.urlopen(url)
-	    doc = minidom.parse(xml)
-	    entries = doc.getElementsByTagName("entry")
+		url = "http://search.twitter.com/search.atom?rpp=" + count + "&q=%s&since_id=%s" % (query, id)
+		xml = urllib.urlopen(url)
+		doc = minidom.parse(xml)
+		entries = doc.getElementsByTagName("entry")
 
-	    if len(entries) > 0:
-		entries.reverse()
-		for e in entries:
-		    title = e.getElementsByTagName("title")[0].firstChild.data
-		    pub = e.getElementsByTagName("published")[0].firstChild.data
-		    id = e.getElementsByTagName("id")[0].firstChild.data.split(":")[2]
-		    name = e.getElementsByTagName("name")[0].firstChild.data.split(" ")[0]
-		    
-		    try:
-				send_output(module, name + ": " + title + " [" + pub + "]")
-	 	    except:
-				print_warning(module, "Couldn't print line because it contains non-ASCII values.")
+		if len(entries) > 0:
+			entries.reverse()
+			#If entries, interate through entries and output results.
+			for e in entries:
+		    		title = e.getElementsByTagName("title")[0].firstChild.data
+		    		pub = e.getElementsByTagName("published")[0].firstChild.data
+		    		id = e.getElementsByTagName("id")[0].firstChild.data.split(":")[2]
+		    		name = e.getElementsByTagName("name")[0].firstChild.data.split(" ")[0]
+		    		#Try output...non-ascii will through an exception.
+		 		try:
+					send_output(module, name + ": " + title + " [" + pub + "]")
+	 	    		except:
+					print_warning(module, "Couldn't print line because it contains non-ASCII values.")
 	    
+		#Set delay should be at least 5 seconds.
 		time.sleep(interval)
 
 #Start the module
